surveyor/requirements.txt
django==4.2.9
django-csp
# django-password-validators
django_bootstrap_icons
crispy-bootstrap5
django-crispy-forms
django-mathfilters
django-tz-detect
folium
influxdb-client
matplotlib
pandas==2.1.4
requests
numpy
gunicorn
pytz
geopandas
networkx
libpysal
rtree
icecream
seaborn
celery
django-celery-beat
django-celery-results
redis
django-redis
pluscodes
pyyaml
